* commit 'eaccccf8297e792adb5423a16a66dacfc8d50c5e': (226 commits)
  VFS-6362 Added bamboos submodule
  VFS-6284 improve descriptions
  VFS-6284 rename luma endpoints associated with storage_sync to storage_import
  VFS-6284 improve descriptions
  VFS-6284 split LUMA tag
  VFS-6284 introduce improvements according to code review
  VFS-6437 Fix rest_model.erl generation
  VFS-6437 Reorder inherited objects for storage models
  VFS-6284 improve definitions and descriptions
  Updated version to 20.02.0-beta4
  Releasing new version 19.02.3
  VFS-6284 add API for fetching/deleting mappings in LUMA DB, add API for setting/getting/deleting mappings in LUMA DB with local feed
  VFS-6284 update configuration of storage: remove insecure option, rename readonly to skipStorageDetection, change lumaEnabled to lumaMode
  Updated Bash codegen docker image
  Release tags
  Bump version to 20.02.0-beta3
  VFS-6279 Fix bad auth definitions in swagger that break the REST-CLI generation
  VFS-6154 Add REST field listing all cluster nodes
  VFS-5682 Allow changing fileMode and dirMode in webdav storage update
  VFS-6105 Remove credentials from join_cluster x-code-samples
  ...

+++ Makefile
@@ -8,10 +8,14 @@ SWAGGER_MARKDOWN_IMAGE      ?= docker.onedata.org/swagger-gitbook:1.4.1
 SWAGGER_COWBOY_SERVER_IMAGE ?= docker.onedata.org/swagger-codegen:2.3.1-cowboy
 SWAGGER_PYTHON_CLIENT_IMAGE ?= docker.onedata.org/swagger-codegen:2.2.2-1b1767e
 SWAGGER_JS_CLIENT_IMAGE     ?= docker.onedata.org/swagger-codegen:VFS-3144
-SWAGGER_BASH_CLIENT_IMAGE   ?= docker.onedata.org/swagger-codegen:ID-2fc8126ac8
+SWAGGER_BASH_CLIENT_IMAGE   ?= docker.onedata.org/swagger-codegen:VFS-6328
 SWAGGER_REDOC_IMAGE         ?= docker.onedata.org/swagger-redoc:1.0.0
 
-BRANCH = $(shell git rev-parse --abbrev-ref HEAD)
+ifndef BRANCH
+	# BRANCH should always be passed in case of git detached state. This is only a fallback.
+	BRANCH = $(shell git rev-parse --abbrev-ref HEAD)
+endif
+
 COMMIT_MESSAGE = $(shell git log -1 --pretty=format:%s)
 
 .PHONY : all swagger.json
@@ -33,7 +37,7 @@ validate: swagger.json
 	fi
 
 cowboy-server: validate
-	docker run --rm -e CHOWNUID=${UID} -v `pwd`:/swagger -t ${SWAGGER_COWBOY_SERVER_IMAGE} generate -i ./swagger.json -l cowboy -o ./generated/cowboy
+	docker run --rm -e CHOWNUID=${UID} -v `pwd`:/swagger -t ${SWAGGER_COWBOY_SERVER_IMAGE} generate -i ./swagger.json -l cowboy -o ./generated/cowboy -DapiFileNameSuffix="_rest_routes"
 	./fix_generated.py
 
 python-client: validate
@@ -46,10 +50,12 @@ javascript-client: validate
 	docker run --rm -e "CHOWNUID=${UID}" -v `pwd`:/swagger -t ${SWAGGER_JS_CLIENT_IMAGE}  generate -i ./swagger.json -l javascript -o ./generated/javascript/ -c ./javascript-config.json
 
 javascript-update-repo: clean javascript-client
+	if [ "${BRANCH}" = "HEAD" ]; then exit 1; fi
 	rm -rf generated/javascript-git
-	git clone ssh://git@git.plgrid.pl:7999/vfs/onepanel-javascript-client.git generated/javascript-git && \
+	git clone ssh://git@git.onedata.org:7999/vfs/onepanel-javascript-client.git generated/javascript-git && \
 	cd generated/javascript-git && \
 	( ( git checkout ${BRANCH} && ( git pull || exit 1 ) ) || git checkout -b ${BRANCH} ) && \
+	git rm -r * && \
 	cp -R ../javascript/* . && \
 	git add -A . && \
 	git config user.email "bamboo@onedata.org" && \
@@ -103,3 +109,7 @@ bash-packages:
         cp generated/bash/onepanel-rest-cli.bash-completion "packages/bash/$$release_branch/";\
     done
 	@git checkout master
+
+submodules:
+	git submodule sync --recursive ${submodule}
+	git submodule update --init --recursive ${submodule}
